+++ src/components/LifecyleProgress/LifecycleStages.tsx
@@ -1,12 +1,17 @@
 import styled, {css} from "styled-components";
 import type {StageColumnComponent, StageBlockComponent, StageIndicatorComponent} from '@models/interfaces/ILifecycleProgressProps';
 
-export const StagesColumn = styled.div<StageColumnComponent>`
-    display: grid;
-    grid-template-columns: ${p => p?.gridTemplateColumn ?? 'repeat(auto-fit, minmax(200px, 1fr))'};
-    background: #efefef;
-    height: ${p => p?.columnsHeight ?? '46px'};
+export const StagesColumn = styled.div.attrs<StageColumnComponent>(() => ({className: 'stage-column'}))<StageColumnComponent>`
+    background: ${p => p?.containerBgColor ?? '#fff'};
+    height: ${p => p?.rightColumnHeight ?? '73px'};
     width: 100%;
+    display: flex;
+    align-items: flex-end;
+    padding-bottom: ${p=> p?.rightColPaddingBot ?? '0'};
+    overflow-x: overlay;
+    * {
+        box-sizing: border-box;
+    }
 `;
 
 const pseudoCss = (type: 'after'| 'before', ...content: string[]) => {
@@ -16,11 +21,16 @@ const pseudoCss = (type: 'after'| 'before', ...content: string[]) => {
         width: 50%;
         position: absolute;
         top: -4px;
+        ${type === 'after' ? 'right: 0;' : 'left: 0;'};
         ${content}
     `;
     if (type === 'after')
-        return css`::after {right: 0; ${baseCss}}`;
-    return css`::before {left: 0; ${baseCss}}`;
+        return css`::after {
+            ${baseCss}
+        }`;
+    return css`::before {
+        ${baseCss}
+    }`;
 }
 
 export const StageBlock = styled.div<StageBlockComponent>`
@@ -28,11 +38,12 @@ export const StageBlock = styled.div<StageBlockComponent>`
     flex-direction: column;
     justify-content: flex-end;
     align-items: center;
-    height: ${p => p?.columnsHeight ?? '46px'};
+    height: ${p => p?.stageHeight ?? '59px'};
     border-top: ${p => p?.stageBorderTop ?? '4px solid #e2e2e2'};
     position: relative;
     background-color: ${p => p?.stageBgColor ?? '#efefef'};
     width: 100%;
+    min-width: ${p => p?.stageMinWidth ?? '200px'};
     ${p => !p?.isFirstColumn && pseudoCss('before', (p?.completed || p?.active) ? `background-color: ${p?.indicatorColor ?? '#00BCF2'}` : '')}
     ${p => !p?.isLastColumn && pseudoCss('after', p?.completed ? `background-color: ${p?.indicatorColor ?? '#00BCF2'}` : '')}
 `;
@@ -43,9 +54,9 @@ const stageIndicatorBgColor = (completed: boolean, defaultColor: string) => {
     return '#fff';
 }
 
-export const StageIndicator = styled.span<StageIndicatorComponent>`
-    width: ${p => (p?.active || (!p?.active && !p?.completed)) ? '20px' : '24px'};
-    height: ${p => (p?.active || (!p?.active && !p?.completed)) ? '20px' : '24px'};
+export const StageIndicator = styled.span.attrs<StageIndicatorComponent>(() => ({className: 'indicator-circle'}))<StageIndicatorComponent>`
+    width: 24px;
+    height: 24px;
     border-radius: 24px;
     background-color: ${p => stageIndicatorBgColor(p?.completed, p?.indicatorColor)};
     border: ${p => (p?.completed) ? 'unset' : (p?.active) ? `4px solid ${p?.indicatorColor ?? '#00BCF2'}` : '4px solid #c4c4c4'};