+++ src/models/interfaces/ILifecycleProgressProps.ts
@@ -1,22 +1,25 @@
 import type {IFocusTrapCalloutProps} from '@fluentui/react/lib/Callout';
+import type {CSSNumberFormat} from '@models/types/UtilityTypes';
 
 export interface IInfoColumn {
     infoColumnBgColor?: string;
     /**@default #fff */
     infoColumnTxtColor?: string;
     /**@default 200px */
-    infoColumnMaxWidth?: string;
+    infoColumnWidth?: CSSNumberFormat;
     infoContent: React.ReactNode;
 }
 
 export interface IStageColumn extends IStageBlock {
-     /**@default `repeat(auto-fit, minmax(200px, 1fr))` */
-    gridTemplateColumn?: string;
     /**@default #00BCF2 */
     indicatorColor?: string;
 }
 
 export interface IStageBlock {
+    /**@default 42px */
+    stageHeight?: CSSNumberFormat;
+    /**@default 200px */
+    stageMinWidth?: CSSNumberFormat;
     stageBgColor?: string;
     stageBorderTop?: string;
 }
@@ -33,11 +36,11 @@ export interface ILifecycleStages<T = any> {
 }
 
 export interface ILifecycleProgressProps<StageData = any> extends IInfoColumn, IStageColumn {
-    /**@default 60px */
-    containerHeight?: string;
     /**@default #00BCF2 */
-    /**@default 46px */
-    columnsHeight?: string;
+    /**@default 59px */
+    leftColumnHeight?: CSSNumberFormat;
+    /**@default 73px */
+    rightColumnHeight?: CSSNumberFormat;
     /**@default #fff */
     containerBgColor?: string;
     stages: ILifecycleStages<StageData>[];
@@ -52,7 +55,8 @@ export interface ILifecycleProgressProps<StageData = any> extends IInfoColumn, I
 
 export type LifecycleCallout = {isVisible: boolean, calloutIdx: number};
 export type ILifecycleProgressRef = {setCallout: React.Dispatch<React.SetStateAction<LifecycleCallout>>};
-export type StageColumnComponent = Pick<ILifecycleProgressProps, 'gridTemplateColumn' | 'columnsHeight'>;
-export type StageBlockComponent = Pick<ILifecycleProgressProps, 'stageBgColor' | 'indicatorColor' | 'stageBorderTop' | 'columnsHeight'> & {isFirstColumn: boolean, isLastColumn: boolean, completed: boolean, active: boolean};
+export type StageColumnComponent = Pick<ILifecycleProgressProps, 'rightColumnHeight' | 'containerBgColor'> & { rightColPaddingBot: CSSNumberFormat };
+export type StageBlockComponent = Pick<ILifecycleProgressProps, 'stageBgColor' | 'indicatorColor' | 'stageBorderTop' | 'rightColumnHeight' | 'stageHeight' | 'stageMinWidth'> & {isFirstColumn: boolean, isLastColumn: boolean, completed: boolean, active: boolean};
 export type StageIndicatorComponent = Pick<IStageColumn, 'indicatorColor'> & {active: boolean, completed: boolean};
-export type InfoColumnComponent = Pick<ILifecycleProgressProps, 'infoColumnBgColor' | 'infoColumnTxtColor' | 'infoColumnMaxWidth' | 'columnsHeight'>
+export type InfoColumnComponent = Pick<ILifecycleProgressProps, 'infoColumnBgColor' | 'infoColumnTxtColor' | 'leftColumnHeight'>
+export type LifeCycleContainer = Pick<ILifecycleProgressProps, 'containerBgColor' | 'infoColumnWidth' | 'stageMinWidth'>;